[WiP] Modify parser to print JSON
This prints the bp files as something which can be parsed as json,
and used to write makefiles.

Use case: Debian Android Tools packaging effort - to be able to create
makefiles for creating debian packaging by parsing the Android.bp
files, instead of having to do it manually.

TODO: Make it entirely conditional, current usage is by 'bpfmt',
so add a '-j' flag to that.